test: fix test-debugger-pid
The current format that prints the process PID before the actual
message.

Adapt the test so it also passes on Windows, that emits a different
message from the rest of platforms.

Move the test to parallel.

PR-URL: https://github.com/nodejs/node/pull/6584
